refactor

+++ config/model.go
@@ -6,7 +6,7 @@ import (
 )
 
 type MODEL struct {
-    ID        uint           `gorm:"primarykey; autoIncrement:10000"` // 主键ID
+    ID        int64          `gorm:"primarykey; autoIncrement:10000"` // 主键ID
     CreatedAt time.Time      // 创建时间
     UpdatedAt time.Time      // 更新时间
     DeletedAt gorm.DeletedAt `gorm:"index" json:"-"` // 删除时间
